# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "selfevals",
    "description": "HTTP bridge for the selfevals evals framework.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/analysis/bundle": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Analysis Bundle",
        "operationId": "analysis_bundle_api_workspaces__workspace_id__experiments__experiment_id__analysis_bundle_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "iteration",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Iteration"
            }
          },
          {
            "name": "all",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include passing traces, not just failures.",
              "default": false,
              "title": "All"
            },
            "description": "Include passing traces, not just failures."
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Analysis Bundle Api Workspaces  Workspace Id  Experiments  Experiment Id  Analysis Bundle Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/experiments/{experiment_id}/analysis/ingest": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analysis Ingest",
        "operationId": "analysis_ingest_api_workspaces__workspace_id__experiments__experiment_id__analysis_ingest_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "experiment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Experiment Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisIngestSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspace_id}/anchor-set": {
      "get": {
        "tags": [
          "anchor-set"
        ],
        "summary": "Anchor Set",
        "operationId": "anchor_set_api_workspaces__workspace_id__anchor_set_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnchorPoint"
                  },
                  "title": "Response Anchor Set Api Workspaces  Workspace Id  Anchor Set Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/clusters": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Failing traces grouped by failure mode (\u00a7J.6)",
        "description": "Groups failing traces by their stable failure-mode slug, ranked by frequency, with capped example `run_id`s for drill-down and title/status enriched from the workspace taxonomy. v1 clusters by the existing taxonomy (cluster \u2261 mode); semantic clustering is future work.",
        "operationId": "clusters_api_workspaces__workspace_id__clusters_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "experiment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment Id"
            }
          },
          {
            "name": "grader",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grader"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 200,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailureClustersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalysisIngestSummaryResponse": {
        "properties": {
          "assignments_applied": {
            "type": "integer",
            "title": "Assignments Applied"
          },
          "created_candidates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Created Candidates"
          },
          "updated_candidates": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Updated Candidates"
          },
          "hypotheses_recorded": {
            "type": "integer",
            "title": "Hypotheses Recorded"
          }
        },
        "type": "object",
        "required": [
          "assignments_applied",
          "hypotheses_recorded"
        ],
        "title": "AnalysisIngestSummaryResponse"
      },
      "AnchorPoint": {
        "properties": {
          "experiment_id": {
            "type": "string",
            "title": "Experiment Id"
          },
          "experiment_name": {
            "type": "string",
            "title": "Experiment Name"
          },
          "iteration": {
            "type": "integer",
            "title": "Iteration"
          },
          "primary_metric_name": {
            "type": "string",
            "title": "Primary Metric Name"
          },
          "primary_metric_value": {
            "type": "number",
            "title": "Primary Metric Value"
          },
          "decision_outcome": {
            "type": "string",
            "title": "Decision Outcome"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "experiment_id",
          "experiment_name",
          "iteration",
          "primary_metric_name",
          "primary_metric_value",
          "decision_outcome",
          "created_at"
        ],
        "title": "AnchorPoint"
      },
...
      "FailureClusterExample": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "FailureClusterExample",
        "description": "A concrete failing trace inside a cluster \u2014 enough to drill straight into\nthe trace viewer (`/traces/{run_id}` resolves both `run_\u2026` and `tr_\u2026` ids)."
      },
      "FailureClusterRow": {
        "properties": {
          "failure_mode": {
            "type": "string",
            "title": "Failure Mode"
          },
          "failure_mode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Mode Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "unknown"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "rate": {
            "type": "number",
            "title": "Rate"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/FailureClusterExample"
            },
            "type": "array",
            "title": "Examples"
          }
        },
        "type": "object",
        "required": [
          "failure_mode",
          "count",
          "rate"
        ],
        "title": "FailureClusterRow",
        "description": "One cluster = one failure mode (\u00a7J.6 v1: cluster \u2261 taxonomy mode).\n\n`failure_mode` is the stable slug carried on every grade; `title`/`status`\nare enriched from the workspace taxonomy when the mode is registered there\n(`status=\"unknown\"` for a mode seen on a grade but not yet in the taxonomy \u2014\na candidate the analysis loop hasn't formalized). `examples` are capped\nsample traces for drill-down, not the full membership."
      },
      "FailureClustersResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "window": {
            "$ref": "#/components/schemas/MetricsWindow"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FailureClusterRow"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "window",
          "total"
        ],
        "title": "FailureClustersResponse",
        "description": "Failing traces grouped by failure mode, ranked by frequency. Same window\nenvelope as the other metrics so the FE filters identically."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MetricsWindow": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          }
        },
        "type": "object",
        "title": "MetricsWindow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}